Flush stdout after writing autosize dimensions
Enabling autosize for a box previously printed the dimensions to stdout if the
box was created on the command line, but due to not flushing the output,
programs could not get dimensions from the box through piping or other
output redirections.

An example script usage of the autosize output has been added to the
`examples` directory.