Add argument to get box size without drawing
Getting the size of a box was difficult because it involved capturing
the output of a background process, therefore having to delay the
running of any processes that wished to process the box size. This
commit adds a new argument, `--bounds`, which causes the program to
print its autosize boundaries and immediately exit.

This commit also presents some small changes in the way arugmnets are
parsed.